+++ libs/idp-server-core-adapter/src/main/java/org/idp/server/core/adapters/datasource/federation/session/command/PostgresqlExecutor.java
@@ -5,21 +5,25 @@ import java.util.List;
 import org.idp.server.basic.datasource.SqlExecutor;
 import org.idp.server.basic.json.JsonConverter;
 import org.idp.server.core.federation.sso.SsoSessionIdentifier;
+import org.idp.server.core.multi_tenancy.tenant.Tenant;
 
 public class PostgresqlExecutor implements SsoSessionCommandSqlExecutor {
 
   JsonConverter jsonConverter = JsonConverter.snakeCaseInstance();
 
   @Override
-  public <T> void insert(SsoSessionIdentifier identifier, T payload) {
+  public <T> void insert(Tenant tenant, SsoSessionIdentifier identifier, T payload) {
     SqlExecutor sqlExecutor = new SqlExecutor();
 
     String sqlTemplate =
         """
                 INSERT INTO federation_sso_session (
                 id,
-                payload)
+                tenant_id,
+                payload
+                )
                 VALUES (
+                ?::uuid,
                 ?::uuid,
                 ?::jsonb
                 )
@@ -30,6 +34,7 @@ public class PostgresqlExecutor implements SsoSessionCommandSqlExecutor {
     String json = jsonConverter.write(payload);
     List<Object> params = new ArrayList<>();
     params.add(identifier.value());
+    params.add(tenant.identifierValue());
     params.add(json);
     params.add(json);
 
@@ -37,18 +42,19 @@ public class PostgresqlExecutor implements SsoSessionCommandSqlExecutor {
   }
 
   @Override
-  public void delete(SsoSessionIdentifier identifier) {
+  public void delete(Tenant tenant, SsoSessionIdentifier identifier) {
     SqlExecutor sqlExecutor = new SqlExecutor();
 
     String sqlTemplate =
         """
                 DELETE FROM federation_sso_session
-                WHERE id = ?::uuid;
+                WHERE id = ?::uuid
+                AND tenant_id = ?::uuid;
                 """;
 
     List<Object> params = new ArrayList<>();
     params.add(identifier.value());
-
+    params.add(tenant.identifierValue());
     sqlExecutor.execute(sqlTemplate, params);
   }
 }
